\name{readThermocamRaw}
\alias{readThermocamRaw}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A function to read a single thermal image
}
\description{
A function to read a single thermal image
}
\usage{
cooccur(file}
\arguments{
  \item{file}{
Full filename of .seq file
}
}
\value{
Returns a matrix with attribute called info, with camera info.
}

\author{
Daniel M Griffith
}

\examples{

# todo
}

\keyword{read}
