\name{stackThermocamRaw}
\alias{stackThermocamRaw}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A function to read a multiple thermal images
}
\description{
A function to read a multiple thermal image
}
\usage{
cooccur(file}
\arguments{
  \item{file}{
Vector of full filenames for .seq files
}
}
\value{
Returns a raster stack
}

\author{
Daniel M Griffith
}

\examples{

# todo
}

\keyword{read}
